+++ Components/header.tsx
@@ -1,7 +1,5 @@
 import { Rubik } from "@next/font/google";
-import Image from "next/image";
 import Link from "next/link";
-import { Component } from "react";
 import styles from "../styles/Home.module.css";
 
 const rubik = Rubik({ subsets: ["latin"] });
