+++ Components/title.tsx
@@ -1,5 +1,4 @@
 import styles from "../styles/Home.module.css"
-
 interface title {
   title: string,
   color: string,
@@ -9,7 +8,7 @@ export default function Title(props: title) {
   return (
     <div className={styles.title_text}>
       <h1 className={styles.header_text}>{props.title}</h1>
-      <div className={styles.header_underline} style={{ backgroundColor: `var(${props.color})` }}></div>
+      <div className={`${styles.header_underline} ${props.color}`}></div>
     </div>
   )
 }
