+++ pages/about.tsx
@@ -1,27 +1,18 @@
-// Fonts
-import { Rubik } from "@next/font/google"
 // Components
 import Head from "next/head"
-import Link from "next/link"
-import Header from "../Components/header"
 import Title from "../Components/title"
 // styles
 import generics from "../styles/generics.module.css"
 import styles from "../styles/Home.module.css"
-
-const rubik = Rubik({ subsets: ['latin'] })
 
 export default function About() {
 	return (
 		<>
 			<Head>
 				<title>About</title>
-				<link rel="icon" href="/favicon.io"></link>
+				<link rel="icon" href="/favicon.ico"></link>
 			</Head>
-			<Header />
-			<main className={`${styles.main} ${rubik.className}`}>
-				<Title title="About" color="--main-blue" />
-			</main>
+			<Title title="About" color={generics.main_blue} />
 		</>
 	)
 }
